% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binary.R
\name{binary}
\alias{binary}
\title{Binary-ize a Matrix}
\usage{
binary(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A binary matrix
}
\description{
Takes a numeric matrix and makes it binary (adjacency).
}
\examples{
x <- matrix(data=c(0,4,3,0,0,3,1,2,3),nrow=3,ncol=3)
binary(x)
}

