% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cbindr.R
\name{cbindr}
\alias{cbindr}
\title{Split df columns}
\usage{
cbindr(df, colts, pattern, newnames)
}
\arguments{
\item{df}{The data frame containing the column to split}

\item{colts}{Column to split}

\item{pattern}{Regex pattern on which to split the column}

\item{newnames}{A list of names to assign to the new columns}
}
\value{
The inputted data frame with both the split column and the new columns
}
\description{
Computes weighted variance for a variable
}
\details{
This function computes the weighted variables for a variable. CURRENTLY NOT WORKING.
}
\examples{
cars$distsp <- paste(cars$dist,'.0',sep='')
cars <- cbindr(cars,'distsp','.',c('orgdist','zero'))
}

