% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/globinfo.R
\name{globinfo}
\alias{globinfo}
\title{GlobInfo: Get Workspace Info}
\usage{
globinfo()
}
\value{
formatted string with global environment info
}
\description{
Tells you about your global_env() neatly.
}
\details{
This package prints out data frames, functions, and other objects in your global environment.
}
\examples{
globinfo()
}

