% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nbl.R
\name{nbl}
\alias{nbl}
\title{Nick Bloom CSV Loader}
\usage{
nbl(x)
}
\arguments{
\item{x}{Path to csv as string}
}
\value{
data frame
}
\description{
Loads CSVs the right way
}
\details{
This package loads CSVs with stringsAsFactor=FALSE, and then sets the names of the resulting data frame to lowercase.
}
\examples{
nbl('/Users/nbloom/Documents/workzone/statistics/nonproject code/bballrankings/bbschools.csv')
}

